+++ tchannel/thrift/client.py
@@ -30,7 +30,7 @@ from tchannel.deprecate import deprecated
 from tchannel.errors import OneWayNotSupportedError
 
 from ..serializer.thrift import ThriftSerializer
-from ..tracing import ClientTracer
+from ..import tracing
 from .reflection import get_service_methods
 
 # Generated clients will use this base class.
@@ -171,7 +171,7 @@ def generate_method(service_module, service_name, method_name):
         for name, value in params.items():
             setattr(call_args, name, value)
 
-        tracer = ClientTracer(channel=self.tchannel)
+        tracer = tracing.ClientTracer(channel=self.tchannel)
         span, headers = tracer.start_span(
             service=service_name, endpoint=method_name, headers={}
         )
@@ -181,7 +181,7 @@ def generate_method(service_module, service_name, method_name):
 
         # Glue for old API.
         if hasattr(self.tchannel, 'request'):
-            tracer.apply_trace_flag(span, self.trace)
+            tracing.apply_trace_flag(span, self.trace, True)
             with span:
                 response = yield self.tchannel.request(
                     hostport=self.hostport, service=self.service
